# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Discography Management API
  description: API for managing artists and albums in a discography system
  version: 1.0.0
  contact:
    name: Marcos Oliveira
    email: marcos@example.com
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/v1/artists:
    get:
      tags:
        - Artists
      summary: Get paginated list of artists
      description: Retrieve a paginated list of all artists
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of artists
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArtistResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean

    post:
      tags:
        - Artists
      summary: Create new artist
      description: Create a new artist with optional profile image upload
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - request
              properties:
                request:
                  type: object
                  description: Artist request data
                  required:
                    - name
                    - genre
                  properties:
                    name:
                      type: string
                      description: Artist name
                      example: "The Beatles"
                    genre:
                      type: string
                      description: Music genre
                      example: "Rock"
                file:
                  type: string
                  format: binary
                  description: Artist profile image (optional)
                  example: "profile-image.jpg"
      responses:
        '201':
          description: Artist created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDTO'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
  /api/v1/artists/{id}:
    get:
      tags:
        - Artists
      summary: Get artist by ID
      description: Retrieve a specific artist by their UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Artist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtistResponseDTO'
        '404':
          description: Artist not found

    put:
      tags:
        - Artists
      summary: Update artist
      description: Update an existing artist's information
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtistRequestDTO'
      responses:
        '204':
          description: Artist updated successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'

    delete:
      tags:
        - Artists
      summary: Delete artist
      description: Delete an artist by their UUID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '204':
          description: Artist deleted successfully
  /api/v1/artists/name/{name}:
    get:
      tags:
        - Artists
      summary: Find artists by name
      description: Search for artists by name with pagination support
      parameters:
        - name: name
          in: path
          required: true
          schema:
            type: string
            description: Artist name to search for
            example: "Beatles"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of matching artists
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArtistResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
  /api/v1/artists/genre/{genre}:
    get:
      tags:
        - Artists
      summary: Find artists by genre
      description: Search for artists by music genre with pagination support
      parameters:
        - name: genre
          in: path
          required: true
          schema:
            type: string
            description: Music genre to search for
            example: "Rock"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of artists in the specified genre
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArtistResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
  /api/v1/artists/album/{id}:
    get:
      tags:
        - Artists
      summary: Find artists by album ID
      description: Find all artists associated with a specific album
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Album UUID
            example: "550e8400-e29b-41d4-a716-446655440002"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of artists associated with the album
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/ArtistResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
  /api/v1/artists/{id}/profileImage:
    put:
      tags:
        - Artists
      summary: Upload artist profile image
      description: Upload or update a profile image for an artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            example: "550e8400-e29b-41d4-a716-446655440000"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Profile image file
                  example: "profile-image.jpg"
      responses:
        '204':
          description: Profile image uploaded successfully
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
        '404':
          description: Artist not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorDetails'
...
  /api/v1/albums/artist/{id}:
    get:
      tags:
        - Albums
      summary: Find albums by artist ID
      description: Find all albums associated with a specific artist
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
            description: Artist UUID
            example: "550e8400-e29b-41d4-a716-446655440000"
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of albums by the specified artist
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: array
                    items:
                      $ref: '#/components/schemas/AlbumResponseDTO'
                  pageable:
                    type: object
                    properties:
                      sort:
                        type: object
                      offset:
                        type: integer
                      pageNumber:
                        type: integer
                      pageSize:
                        type: integer
                      paged:
                        type: boolean
                      unpaged:
                        type: boolean
                  last:
                    type: boolean
                  totalPages:
                    type: integer
                  totalElements:
                    type: integer
                  size:
                    type: integer
                  number:
                    type: integer
                  sort:
                    type: object
                    properties:
                      empty:
                        type: boolean
                      sorted:
                        type: boolean
                      unsorted:
                        type: boolean
                  first:
                    type: boolean
                  numberOfElements:
                    type: integer
                  empty:
                    type: boolean
...
components:
  schemas:
    ArtistRequestDTO:
      type: object
      required:
        - name
        - genre
      properties:
        name:
          type: string
          description: Artist name
          example: "The Beatles"
          minLength: 1
          pattern: "^[a-zA-Z0-9\\s\\-\\'\\.]+$"
        genre:
          type: string
          description: Music genre
          example: "Rock"
          minLength: 1
          pattern: "^[a-zA-Z0-9\\s\\-\\'\\.]+$"
    ArtistResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          example: "The Beatles"
        genre:
          type: string
          example: "Rock"
        profileImageUrl:
          type: string
          example: "http://localhost:8080/api/v1/artists/550e8400-e29b-41d4-a716-446655440000/profile-image"
        albumTitles:
          type: array
          items:
            type: string
          example: ["Abbey Road", "Let It Be", "Revolver"]
          minItems: 0
          maxItems: 100
...
    AlbumResponseDTO:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440002"
        title:
          type: string
          example: "Abbey Road"
        releaseDate:
          type: string
          format: date
          example: "1969-09-26"
        coverUrl:
          type: string
          example: "http://localhost:8080/api/v1/albums/550e8400-e29b-41d4-a716-446655440002/cover"
        artistNames:
          type: array
          items:
            type: string
          example: ["The Beatles"]
          minItems: 1
          maxItems: 10
    ErrorDetails:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2024-01-26T10:30:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Artist not found with id: 550e8400-e29b-41d4-a716-446655440000"
        path:
          type: string
          example: "/api/v1/artists/550e8400-e29b-41d4-a716-446655440000"
      required:
        - timestamp
        - status
        - error
        - message
        - path
...